requirements.txt
# ComfyUI 控制器依赖

websocket-client==1.7.0
requests==2.31.0
